feat: expose custom metrics for tenants status
closes #451

This commit adds pkg/stats that defines prometheus counters for  active and
cordoned tenants.

These metrics are exposed to controller-runtime. So they are visible on the
controller metrics endpoint.

This was tested manually , after creating a new tenant the metrics showed
up on the metrics service endpoint

```
capsule_tenants_status_active_count 1
```